+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 
 ### Changed
 - Move entityType and ClientRegistrationType to OpenId Federation general settings
+- Common Metadata publish only to Federation Entity
 
 
 ## [22.0.13-1.22] - 2025-09-08